System display working properly again

+++ Assets/_EvEMap/Scripts/Data/Position.cs
@@ -3,8 +3,8 @@
 namespace _ProjectEvE.Scripts.Data {
     [Serializable]
     public class Position {
-        public long x;
-        public long y;
-        public long z;
+        public double x;
+        public double y;
+        public double z;
     }
 }
